Add CusomtException class in new namespace
Introduced a new namespace `nopCommerceReplicatorServices.nopCommerce` and added a `[Serializable]` internal class `CusomtException` with three constructors: a parameterless constructor, a constructor with a nullable string `message`, and a constructor with a nullable string `message` and a nullable `Exception` object `innerException`. This class is designed for handling exceptions specific to the nopCommerce replicator services.