Add UpdateProductInventoryAsync method to IProduct interface
A new method `UpdateProductInventoryAsync` has been added to the
`IProduct` interface within the `nopCommerceReplicatorServices.Actions`
namespace. This method updates the inventory of a product in
nopCommerce asynchronously. It takes three parameters: `productId`
(the ID of the product from the external service), `productExternal`
(the external product source data), and `setService` (the chosen
service). The method returns an `HttpResponseMessage` and throws a
`CustomException` if the product is not found.