Broaden API error handling to include 400+ status codes
The condition for checking the status code of the API response has been modified. Previously, the code checked if the status code was not `HttpStatusCode.OK`. Now, it checks if the status code is greater than or equal to `HttpStatusCode.BadRequest`. This change broadens the range of status codes that will trigger the custom exception, ensuring that any client or server error (status codes 400 and above) will result in an exception being thrown.