Add UnreplicatedDataException for data replication issues
Introduced a new exception class `UnreplicatedDataException` in the
`nopCommerceReplicatorServices.Exceptions` namespace. This class
inherits from `CustomException` and is intended to be thrown when
data has not been replicated yet. The exception message provides
guidance on checking the DataBinding database and manually adding
data if necessary. This addition aids in managing and debugging
data replication issues in the nopCommerce replication services.