Add option to replicate product prices by product ID
A new command-line option `repProducPricetIdOption` has been added to allow specifying a product ID for price replication from an external service. This option is now included in the root command options and passed to the command handler. Additionally, logic has been implemented to handle the replication process by invoking `ReplicateProductPriceAsync` when the option is specified and greater than 0.