Generalize product ID binding to use default service
Updated the logic for retrieving product ID bindings to use a
generic service cast to `(Service)0` instead of the specific
`SubiektGT` service. This change allows the main service to
handle product addition and linking, while other external
services are limited to updating data. Updated comments to
reflect this new approach.